Move libs to node folder and rename files
