+++ nodes/common/event_constants.js
