+++ nodes/receiver.js
@@ -27,8 +27,8 @@ module.exports = function(RED)
     const STATUS_UNCONFIGURED = {fill: "yellow", shape: "dot", text: "receiver.status.unconfigured"};
     const STATUS_CONNECTING   = {fill: "grey",   shape: "dot", text: "receiver.status.connecting"  };
 
-    const APIFilter = require("../libs/sony-api-filter");
-    const Events = require("../libs/sony-event-constants.js");
+    const APIFilter = require("./common/api_filter.js");
+    const Events = require("./common/event_constants.js");
 
 
     function SonyAudioReceiverNode(config)
